props 타입 수정
